+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"26.05.2026 11:18"
-#define NE_VERSION   L"2026.05.26.11"
+#define NE_PUBLISHED L"26.05.2026 12:57"
+#define NE_VERSION   L"2026.05.26.12"